Refactor after creating code summaries

+++ tgen/core/trace_output/trace_prediction_output.py
@@ -17,7 +17,7 @@ class TracePredictionEntry(TypedDict):
     source: str
     target: str
     score: Optional[float]
-    label: Optional[float]
+    label: Optional[int]
     explanation: Optional[str]
